requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
python-multipart
pillow